Found the bug in the prediction code, we should get a good result on the Kaggle.
I've also split the training code and the test code.

We can now also resume training from a previous model state.

I've saved the model "epoch22.pt", so we can start further training with a pre-trained model.

One drawback: We need to call manually "predict.py" to get the predictions.

There is also a new prediction file that could be submitted as is.